image-process/requirements.txt
pytest
pytest-cov
grpcio
grpcio-tools
pillow
pyzbar
pyjwt
python-dotenv